requirements.txt
httpx>=0.27,<1
pydantic-settings>=2,<3
python-dotenv>=1,<2
cryptography>=42,<45
eth-keys>=0.6,<1
eth-hash[pycryptodome]>=0.7,<1
eth-account>=0.13,<1
web3>=7,<8
miniupnpc>=2.2,<3
pywebview>=5,<6
rich>=13,<15
pystray>=0.19,<1; sys_platform == "win32"
Pillow>=10,<12; sys_platform == "win32"
